+++ resolver/filtering_resolver_test.go
@@ -14,7 +14,7 @@ var _ = Describe("FilteringResolver", func() {
 	var (
 		sut        *FilteringResolver
 		sutConfig  config.FilteringConfig
-		m          *MockResolver
+		m          *mockResolver
 		mockAnswer *dns.Msg
 	)
 
@@ -24,7 +24,7 @@ var _ = Describe("FilteringResolver", func() {
 
 	JustBeforeEach(func() {
 		sut = NewFilteringResolver(sutConfig).(*FilteringResolver)
-		m = &MockResolver{}
+		m = &mockResolver{}
 		m.On("Resolve", mock.Anything).Return(&Response{Res: mockAnswer}, nil)
 		sut.Next(m)
 	})
