+++ resolver/fqdn_only_resolver_test.go
@@ -14,7 +14,7 @@ var _ = Describe("FqdnOnlyResolver", func() {
 	var (
 		sut        *FqdnOnlyResolver
 		sutConfig  config.Config
-		m          *MockResolver
+		m          *mockResolver
 		mockAnswer *dns.Msg
 	)
 
@@ -24,7 +24,7 @@ var _ = Describe("FqdnOnlyResolver", func() {
 
 	JustBeforeEach(func() {
 		sut = NewFqdnOnlyResolver(sutConfig).(*FqdnOnlyResolver)
-		m = &MockResolver{}
+		m = &mockResolver{}
 		m.On("Resolve", mock.Anything).Return(&Response{Res: mockAnswer}, nil)
 		sut.Next(m)
 	})
