+++ resolver/mock_udp_upstream_server.go
@@ -1,138 +1,14 @@
 package resolver
 
 import (
-	"io"
 	"net"
-	"net/http"
-	"net/http/httptest"
 	"strings"
 	"sync/atomic"
 
 	"github.com/0xERR0R/blocky/config"
 	"github.com/0xERR0R/blocky/util"
-
-	"github.com/0xERR0R/blocky/model"
-
 	"github.com/miekg/dns"
-	"github.com/stretchr/testify/mock"
 )
-
-type MockResolver struct {
-	mock.Mock
-	NextResolver
-
-	ResolveFn  func(req *model.Request) (*model.Response, error)
-	ResponseFn func(req *dns.Msg) *dns.Msg
-	AnswerFn   func(t uint16, qName string) *dns.Msg
-}
-
-func (r *MockResolver) Configuration() []string {
-	args := r.Called()
-
-	return args.Get(0).([]string)
-}
-
-func (r *MockResolver) Resolve(req *model.Request) (*model.Response, error) {
-	args := r.Called(req)
-
-	if r.ResolveFn != nil {
-		return r.ResolveFn(req)
-	}
-
-	if r.ResponseFn != nil {
-		return &model.Response{
-			Res:    r.ResponseFn(req.Req),
-			Reason: "",
-			RType:  model.ResponseTypeRESOLVED,
-		}, nil
-	}
-
-	if r.AnswerFn != nil {
-		for _, question := range req.Req.Question {
-			answer := r.AnswerFn(question.Qtype, question.Name)
-			if answer != nil {
-				return &model.Response{
-					Res:    answer,
-					Reason: "",
-					RType:  model.ResponseTypeRESOLVED,
-				}, nil
-			}
-		}
-
-		response := new(dns.Msg)
-		response.SetRcode(req.Req, dns.RcodeBadName)
-
-		return &model.Response{
-			Res:    response,
-			Reason: "",
-			RType:  model.ResponseTypeRESOLVED,
-		}, nil
-	}
-
-	resp, ok := args.Get(0).(*model.Response)
-
-	if ok {
-		return resp, args.Error(1)
-	}
-
-	return nil, args.Error(1)
-}
-
-// TestBootstrap creates a mock Bootstrap
-func TestBootstrap(response *dns.Msg) *Bootstrap {
-	bootstrapUpstream := &MockResolver{}
-
-	b, err := NewBootstrap(&config.Config{})
-	util.FatalOnError("can't create bootstrap", err)
-
-	b.resolver = bootstrapUpstream
-	b.upstream = bootstrapUpstream
-
-	if response != nil {
-		bootstrapUpstream.
-			On("Resolve", mock.Anything).
-			Return(&model.Response{Res: response}, nil)
-	}
-
-	return b
-}
-
-// TestDOHUpstream creates a mock DoH Upstream
-func TestDOHUpstream(fn func(request *dns.Msg) (response *dns.Msg),
-	reqFn ...func(w http.ResponseWriter)) config.Upstream {
-	server := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		body, err := io.ReadAll(r.Body)
-
-		util.FatalOnError("can't read request: ", err)
-
-		msg := new(dns.Msg)
-		err = msg.Unpack(body)
-		util.FatalOnError("can't deserialize message: ", err)
-
-		response := fn(msg)
-		response.SetReply(msg)
-
-		b, err := response.Pack()
-
-		util.FatalOnError("can't serialize message: ", err)
-
-		w.Header().Set("content-type", "application/dns-message")
-
-		for _, f := range reqFn {
-			if f != nil {
-				f(w)
-			}
-		}
-		_, err = w.Write(b)
-
-		util.FatalOnError("can't write response: ", err)
-	}))
-	upstream, err := config.ParseUpstream(server.URL)
-
-	util.FatalOnError("can't resolve address: ", err)
-
-	return upstream
-}
 
 type MockUDPUpstreamServer struct {
 	callCount int32
@@ -196,7 +72,7 @@ func (t *MockUDPUpstreamServer) Close() {
 	}
 }
 
-func CreateConnection() *net.UDPConn {
+func createConnection() *net.UDPConn {
 	a, err := net.ResolveUDPAddr("udp4", ":0")
 	util.FatalOnError("can't resolve address: ", err)
 
@@ -207,7 +83,7 @@ func CreateConnection() *net.UDPConn {
 }
 
 func (t *MockUDPUpstreamServer) Start() config.Upstream {
-	ln := CreateConnection()
+	ln := createConnection()
 
 	ladr := ln.LocalAddr().String()
 	host := strings.Split(ladr, ":")[0]
