+++ resolver/sudn_resolver_test.go
@@ -14,7 +14,7 @@ import (
 var _ = Describe("SudnResolver", Label("sudnResolver"), func() {
 	var (
 		sut        *SpecialUseDomainNamesResolver
-		m          *MockResolver
+		m          *mockResolver
 		mockAnswer *dns.Msg
 
 		err  error
@@ -25,7 +25,7 @@ var _ = Describe("SudnResolver", Label("sudnResolver"), func() {
 		mockAnswer, err = util.NewMsgWithAnswer("example.com.", 300, dns.Type(dns.TypeA), "123.145.123.145")
 		Expect(err).Should(Succeed())
 
-		m = &MockResolver{}
+		m = &mockResolver{}
 		m.On("Resolve", mock.Anything).Return(&Response{Res: mockAnswer}, nil)
 
 		sut = NewSpecialUseDomainNamesResolver().(*SpecialUseDomainNamesResolver)
